Optimized RDF::Cassandra::Repository#empty?.
Since Cassandra row keys stick around even after they've been deleted, #empty? can be an expensive operation in the worst case where the repository contains mostly or only deleted resources. If you're satisfied with a probabilistic answer as to whether the repository is empty, pass in an integer value to the new :sample option to indicate the maximum number of resources (i.e. row keys) to be examined.